+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Integration between `MST` and Symja `IExpr`
 - Complex power
 - Separate methods for UInt, Int and Number powers. NaN safety.
+- Tensorflow prototype
 
 ### Changed
 - Exponential operations merged with hyperbolic functions